Extract `CuttlefishConfig::InstanceSpecific::GuestOs`
Bug: b/424305767
Test: bazel test '//...' --test_summary=terse
